requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.1
numpy==2.3.2
scikit-learn==1.7.1
xgboost==3.0.2
scipy==1.16.1
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.2.0
openpyxl==3.1.5
influxdb-client==1.49.0
python-dotenv==1.0.0
PyYAML==6.0.1
reactivex==4.0.4
httpx==0.28.1
pandas-stubs==2.3.0.250703
plotly-stubs==0.0.6
types-openpyxl==3.1.5.20250602
types-pytz==2025.2.0.20250516
types-PyYAML==6.0.12.20250516
